+++ Userland/SampleCodeModule/include/phylo.h
@@ -1,1 +1,1 @@
-void phylo(int argc, char * argv[]);
+void phylo(uint64_t argc, char * argv[]);
